requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
azure-core==1.34.0
azure-search-documents==11.6.0
numpy==1.26.4
pypdf==4.3.1
sentence-transformers==2.7.0